BUG: Rename files with similar function, python module, and cython extension names.
