+++ skimage/feature/__init__.py
@@ -1,5 +1,5 @@
 from ._hog import hog
-from .greycomatrix import greycomatrix, greycoprops
+from ._greycomatrix import greycomatrix, greycoprops
 from .peak import peak_local_max
 from ._harris import harris
 from .template import match_template
