+++ skimage/morphology/__init__.py
@@ -2,5 +2,5 @@ from .grey import *
 from .selem import *
 from .ccomp import label
 from .watershed import watershed, is_local_maximum
-from .skeletonize import skeletonize, medial_axis
+from ._skeletonize import skeletonize, medial_axis
 from .convex_hull import convex_hull_image
