+++ skimage/morphology/_skeletonize_cy.pyx
